refactor: infrastructure -> domain 모듈 이동
