+++ invitation-domain/src/main/java/com/nailseong/invitation/member/domain/MemberRepository.java
@@ -1,0 +1,9 @@
+package com.nailseong.invitation.member.domain;
+
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    Optional<Member> findByUsername(String username);
+}
