--- invitation-infrastructure/src/main/java/com/nailseong/invitation/member/MemberRepository.java
@@ -1,9 +1,0 @@
-package com.nailseong.invitation.member;
-
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
-
-    Optional<MemberEntity> findByUsername(String username);
-}
